+++ hslua-core/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 [1]: https://pvp.haskell.org
 
+## hslua-core 2.0.0
+
+Release pending.
+
+- Added new module HsLua.Core.Userdata. It contains thin wrappers
+  around the functions available for creating
+  Haskell-value-wrapping userdata objects.
+
 ## hslua-core 1.0.0
 
 Released 2021-02-27.